requirements.txt
pyperclip
pyinstaller
pillow